# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CO-Finance - Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts-payable:
    get:
      operationId: listAccountsPayable
      tags: [accounts-payable]
      summary: List accounts payable
      parameters:
        - name: profileId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, paid, overdue, cancelled]
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Accounts payable list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountPayable"
    post:
      operationId: createAccountPayable
      tags: [accounts-payable]
      summary: Create an account payable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountPayableBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPayable"
  /accounts-payable/{id}:
    get:
      operationId: getAccountPayable
      tags: [accounts-payable]
      summary: Get an account payable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account payable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPayable"
        "404":
          description: Not found
    patch:
      operationId: updateAccountPayable
      tags: [accounts-payable]
      summary: Update an account payable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountPayableBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPayable"
    delete:
      operationId: deleteAccountPayable
      tags: [accounts-payable]
      summary: Delete an account payable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /accounts-payable/{id}/recurrence-info:
    get:
      operationId: getAccountPayableRecurrenceInfo
      tags: [accounts-payable]
      summary: Returns information about the recurrence series this payable belongs to (used to warn before deletion)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recurrence info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurrenceInfo"
  /accounts-payable/{id}/pay:
    post:
      operationId: payAccountPayable
      tags: [accounts-payable]
      summary: Mark an account payable as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayAccountBody"
      responses:
        "200":
          description: Paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPayable"

  # ── Accounts Receivable ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    AccountPayable:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [open, paid, overdue, cancelled]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        commitmentTypeId:
          type: ["integer", "null"]
        commitmentTypeName:
          type: ["string", "null"]
        personId:
          type: ["integer", "null"]
        personName:
          type: ["string", "null"]
        bankAccountId:
          type: ["integer", "null"]
        invoiceId:
          type: ["integer", "null"]
        paidAt:
          type: ["string", "null"]
        paidAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        recurrent:
          type: boolean
        natureza:
          type: ["string", "null"]
          enum: [manual, card, financing, null]
          description: Nature of the payable - manual entry, credit card invoice, or financing
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, profileId, description, amount, dueDate, status, recurrent, createdAt, updatedAt]
    CreateAccountPayableBody:
      type: object
      properties:
        profileId:
          type: integer
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        categoryId:
          type: ["integer", "null"]
        commitmentTypeId:
          type: ["integer", "null"]
        personId:
          type: ["integer", "null"]
        bankAccountId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        recurrent:
          type: boolean
        natureza:
          type: ["string", "null"]
          enum: [manual, card, financing, null]
        repeatMonths:
          type: integer
          minimum: 1
          maximum: 60
          description: Se informado e > 1, gera cópias mensais até totalizar repeatMonths registros
      required: [profileId, description, amount, dueDate]
    UpdateAccountPayableBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [open, paid, overdue, cancelled]
        categoryId:
          type: ["integer", "null"]
        commitmentTypeId:
          type: ["integer", "null"]
        personId:
          type: ["integer", "null"]
        bankAccountId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        recurrent:
          type: boolean
        natureza:
          type: ["string", "null"]
          enum: [manual, card, financing, null]
    RecurrenceInfo:
      type: object
      properties:
        recurrenceGroupId:
          type: ["string", "null"]
        futureCount:
          type: integer
          description: Quantidade de parcelas futuras em aberto que serão removidas junto desta (não inclui a própria parcela)
      required: [recurrenceGroupId, futureCount]
    PayAccountBody:
      type: object
      properties:
        paidAt:
          type: string
          format: date
        paidAmount:
          type: number
        bankAccountId:
          type: integer
          description: Conta bancária da qual o pagamento será debitado (obrigatório)
      required: [paidAt, paidAmount, bankAccountId]

    # ── Account Receivable ───────────────────────────────────────────────────
...